// src/infrastructure/interfaces/auth.response.ts
// Generated by https://quicktype.io

import {User} from "../../domain/entities/user";

export interface AuthResponse {
  id: string;
  email: string;
  fullName: string;
  isActive: boolean;
  roles: string[];
  token: string;
}

export interface AuthDataResult {
  user?: User | undefined;
  token?: string | undefined;
  ok: boolean;
  message?: string[];
}